Restack onto EditorSanitizer, and cover the policy where it is now consumed
Rebased onto #1231's reshaped tip. The src/ change needed nothing: buildConfig
still returns the whole config and its caller still hands that object, unchanged,
to DOMPurify.sanitize — it is just EditorSanitizer's constructor and #config now
rather than the helper's WeakMap. TRUSTED_TYPES_POLICY rides along either way.

The tests did need moving. Rename detection followed the helper test file to
unit/editor/sanitizer.test.js, which is now about EditorSanitizer's isolation
from the host singleton — not about config/dom_purify, which is what these
exercise and which the refactor did not touch. They join the other
dom_purify_* files under unit/sanitization instead.

Two new cases, because "clean merge" was not evidence that the policy still
arrives.

An editor's sanitizer now sanitizes through a stub policy, and the call is
asserted to have gone through the one named `lexxy`. That crosses the
constructor/#config boundary the refactor introduced, which assertions on
buildConfig's return value do not.

And the same for a sanitizer with no allowlist, which is the fallback #1231 made
deterministic. It has no allowlist to derive keys from and still has to carry the
policy: without it DOMPurify asks the browser for one of its own on our instance,
under the `dompurify` name, so an unregistered consumer sanitizing first would
cause the exact collision this PR exists to prevent. That is why the key is
assigned in buildConfig rather than beside the allowlist — and the case is
mutation-checked by moving it beside the allowlist, which fails it and nothing
else.

The name matters to both assertions. With no policy in the config DOMPurify asks
for one of its own, so a stub recording every createHTML sees the call either way
— checked, and it did pass against the mutation until the recorded name was part
of the expectation. Dropping the TRUSTED_TYPES_POLICY assignment fails these and
the buildConfig case together.

Also drops the template literal from the first dynamic import. The interpolated
value was always 0, and vite's dynamic-import-vars plugin warned on every run.
