+++ test/javascript/unit/editor/sanitizer.test.js
@@ -116,78 +116,3 @@ test("an editor denying attachment content strips it when another editor allows 
   // Only the attribute is refused — the element itself is still allowed here.
   expect(result).toContain("action-text-attachment")
 })
-
-// Every DOMPurify instance claims a policy named `dompurify` under Trusted Types,
-// and TT rejects a duplicate — so on a page with a host sanitizer, ours would get
-// none, and an unsigned instance throws at DOMParser rather than degrading.
-// Ours is created under its own name and handed over, so both work.
-test("uses its own Trusted Types policy name", async () => {
-  const created = []
-  const previous = globalThis.trustedTypes
-
-  globalThis.trustedTypes = {
-    createPolicy(name, rules) {
-      if (created.includes(name)) throw new TypeError(`Policy "${name}" already exists`)
-      created.push(name)
-      return { createHTML: rules.createHTML, createScriptURL: rules.createScriptURL }
-    },
-    getAttributeType: () => null
-  }
-
-  try {
-    // Re-import so the module-scope policy is created against the stub.
-    const fresh = await import(`src/config/dom_purify?tt=${created.length}`)
-    const config = fresh.buildConfig([ "b" ])
-
-    expect(created).toContain("lexxy")
-    expect(created).not.toContain("dompurify")
-    expect(config.TRUSTED_TYPES_POLICY).toBeTruthy()
-  } finally {
-    globalThis.trustedTypes = previous
-  }
-})
-
-// The other two shapes the policy can land in. Creating a policy the CSP has not
-// allowlisted throws, and that throw is caught — so the interesting assertion is
-// not just "no crash" but that the config comes back with the key *absent*.
-// DOMPurify reads cfg.TRUSTED_TYPES_POLICY and validates its shape, so a present
-// -but-undefined key is not the same as an omitted one.
-test("falls back to no policy when the CSP refuses the name", async () => {
-  const previous = globalThis.trustedTypes
-  let asked = false
-
-  globalThis.trustedTypes = {
-    createPolicy(name) {
-      asked = true
-      throw new TypeError(`Refused to create a TrustedTypePolicy named '${name}'`)
-    },
-    getAttributeType: () => null
-  }
-
-  try {
-    const fresh = await import("src/config/dom_purify?tt=refused")
-    const config = fresh.buildConfig([ "b" ])
-
-    expect(asked, "should have attempted to create its own policy").toBe(true)
-    expect(config).not.toHaveProperty("TRUSTED_TYPES_POLICY")
-    // Still sanitizes — this is the position Lexxy was in before it asked at all.
-    expect(fresh.DOMPurify.sanitize("<b>hi</b><script>x()</script>", config)).toBe("<b>hi</b>")
-  } finally {
-    globalThis.trustedTypes = previous
-  }
-})
-
-test("asks for no policy in a browser without Trusted Types", async () => {
-  const previous = globalThis.trustedTypes
-  delete globalThis.trustedTypes
-
-  try {
-    const fresh = await import("src/config/dom_purify?tt=absent")
-    const config = fresh.buildConfig([ "b" ])
-
-    expect(config).not.toHaveProperty("TRUSTED_TYPES_POLICY")
-    expect(fresh.DOMPurify.sanitize("<b>hi</b><script>x()</script>", config)).toBe("<b>hi</b>")
-  } finally {
-    globalThis.trustedTypes = previous
-  }
-})
